broadly separate CLI and GUI code
