+++ src/cli/azure/azureTools.ts
@@ -1,13 +1,13 @@
 import { window, env, Uri } from 'vscode';
-import { globalState, telemetry } from '../extension';
-import { ClusterMetadata } from '../globalState';
+import { globalState, telemetry } from '../../extension';
+import { ClusterMetadata } from '../../globalState';
 import { kubernetesTools } from '../kubernetes/kubernetesTools';
-import { ClusterProvider, ConfigMap, knownClusterProviders } from '../types/kubernetes/kubernetesTypes';
-import { shell, shellCodeError, ShellResult } from '../shell/shell';
-import { TelemetryErrorEventNames } from '../telemetry';
-import { parseJson } from '../utils/jsonUtils';
-import { getCurrentClusterInfo, refreshAllTreeViews } from '../treeviews/treeViews';
-import { failed } from '../types/errorable';
+import { ClusterProvider, ConfigMap, knownClusterProviders } from '../../types/kubernetes/kubernetesTypes';
+import { shell, shellCodeError, ShellResult } from '../../shell/shell';
+import { TelemetryErrorEventNames } from '../../types/telemetryEventNames';
+import { parseJson } from '../../utils/jsonUtils';
+import { getCurrentClusterInfo, refreshAllTreeViews } from '../../treeviews/treeViews';
+import { failed } from '../../types/errorable';
 import { fluxTools } from '../flux/fluxTools';
 import { getAzureMetadata } from './getAzureMetadata';
 import { checkAzurePrerequisites } from './azurePrereqs';
