+++ src/cli/azure/getAzureMetadata.ts
@@ -2,9 +2,9 @@ import { window } from 'vscode';
 import safesh from 'shell-escape-tag';
 import { QuickPickItem } from 'vscode';
 import { kubernetesTools } from '../kubernetes/kubernetesTools';
-import { ClusterProvider, ConfigMap } from '../types/kubernetes/kubernetesTypes';
-import { shell, ShellResult } from '../shell/shell';
-import { parseJson } from '../utils/jsonUtils';
+import { ClusterProvider, ConfigMap } from '../../types/kubernetes/kubernetesTypes';
+import { shell, ShellResult } from '../../shell/shell';
+import { parseJson } from '../../utils/jsonUtils';
 import { AzureClusterProvider, AzureConstants } from './azureTools';
 
 export interface AzureMetadata {
