+++ src/cli/checkVersions.ts
@@ -1,11 +1,10 @@
 import { commands, Uri, window } from 'vscode';
-import { installFluxCli } from './commands/installFluxCli';
-import { enabledWGE, telemetry } from './extension';
-import { shell, shellCodeError } from './shell/shell';
-import { TelemetryErrorEventNames } from './telemetry';
-import { Errorable, failed } from './types/errorable';
-import { CommandId } from './types/extensionIds';
-import { parseJson } from './utils/jsonUtils';
+import { enabledWGE, telemetry } from '../extension';
+import { shell, shellCodeError } from '../shell/shell';
+import { TelemetryErrorEventNames } from '../types/telemetryEventNames';
+import { Errorable, failed } from '../types/errorable';
+import { CommandId } from '../types/extensionIds';
+import { parseJson } from '../utils/jsonUtils';
 
 interface KubectlVersion {
 	major: string;
@@ -93,34 +92,6 @@ export async function getFluxVersion(): Promise<Errorable<string>> {
 			succeeded: false,
 			error: [shellCodeError(fluxVersionShellResult)],
 		};
-	}
-}
-
-/**
- * Show notification with button to install flux
- * (only when flux was not found).
- */
-export async function promptToInstallFlux(): Promise<Errorable<null>> {
-	const version = await getFluxVersion();
-	if (failed(version)) {
-		showInstallFluxNotification();
-		return {
-			succeeded: false,
-			error: ['Flux not found'],
-		};
-	} else {
-		return {
-			succeeded: true,
-			result: null,
-		};
-	}
-}
-
-async function showInstallFluxNotification() {
-	const installButton = 'Install Flux';
-	const pressedButton = await window.showErrorMessage('Please install flux CLI to use GitOps Tools.', installButton);
-	if (pressedButton === installButton) {
-		installFluxCli();
 	}
 }
