+++ src/cli/git/gitInfo.ts
@@ -2,11 +2,11 @@ import path from 'path';
 
 import gitUrlParse from 'git-url-parse';
 import { workspace, window } from 'vscode';
-import { makeSSHUrlFromGitUrl } from '../commands/createSource';
-import { checkGitVersion } from '../install';
+import { makeSSHUrlFromGitUrl } from '../../commands/createSource';
+import { checkGitVersion } from '../checkVersions';
 import { kubernetesTools } from '../kubernetes/kubernetesTools';
-import { GitRepository } from '../types/flux/gitRepository';
-import { shell } from '../shell/shell';
+import { GitRepository } from '../../types/flux/gitRepository';
+import { shell } from '../../shell/shell';
 
 
 export interface GitInfo {
