+++ src/cli/kubernetes/kubernetesTools.ts
@@ -3,23 +3,23 @@ import safesh from 'shell-escape-tag';
 import { Uri, window } from 'vscode';
 import * as kubernetes from 'vscode-kubernetes-tools-api';
 import { AzureConstants } from '../azure/azureTools';
-import { globalState, setVSCodeContext, telemetry } from '../extension';
-import { output } from '../shell/output';
-import { shellCodeError } from '../shell/shell';
-import { TelemetryErrorEventNames } from '../telemetry';
-import { Errorable, failed, succeeded } from '../types/errorable';
-import { ContextId } from '../types/extensionIds';
-import { BucketResult } from '../types/flux/bucket';
-import { GitOpsTemplateResult } from '../types/flux/gitOpsTemplate';
-import { GitRepositoryResult } from '../types/flux/gitRepository';
-import { HelmReleaseResult } from '../types/flux/helmRelease';
-import { HelmRepositoryResult } from '../types/flux/helmRepository';
-import { KustomizeResult } from '../types/flux/kustomize';
-import { OCIRepositoryResult } from '../types/flux/ociRepository';
-import { KubernetesConfig, KubernetesContextWithCluster } from '../types/kubernetes/kubernetesConfig';
-import { KubernetesFileSchemes } from '../types/kubernetes/kubernetesFileSchemes';
-import { ClusterProvider, ConfigMap, DeploymentResult, NamespaceResult, NodeResult, PodResult } from '../types/kubernetes/kubernetesTypes';
-import { parseJson } from '../utils/jsonUtils';
+import { globalState, setVSCodeContext, telemetry } from '../../extension';
+import { output } from '../../shell/output';
+import { shellCodeError } from '../../shell/shell';
+import { TelemetryErrorEventNames } from '../../types/telemetryEventNames';
+import { Errorable, failed, succeeded } from '../../types/errorable';
+import { ContextId } from '../../types/extensionIds';
+import { BucketResult } from '../../types/flux/bucket';
+import { GitOpsTemplateResult } from '../../types/flux/gitOpsTemplate';
+import { GitRepositoryResult } from '../../types/flux/gitRepository';
+import { HelmReleaseResult } from '../../types/flux/helmRelease';
+import { HelmRepositoryResult } from '../../types/flux/helmRepository';
+import { KustomizeResult } from '../../types/flux/kustomize';
+import { OCIRepositoryResult } from '../../types/flux/ociRepository';
+import { KubernetesConfig, KubernetesContextWithCluster } from '../../types/kubernetes/kubernetesConfig';
+import { KubernetesFileSchemes } from '../../types/kubernetes/kubernetesFileSchemes';
+import { ClusterProvider, ConfigMap, DeploymentResult, NamespaceResult, NodeResult, PodResult } from '../../types/kubernetes/kubernetesTypes';
+import { parseJson } from '../../utils/jsonUtils';
 
 export let currentContextName: string;
