+++ src/commands/commands.ts
@@ -1,7 +1,7 @@
 import { commands, Disposable, ExtensionContext, Uri, window } from 'vscode';
 import { telemetry } from '../extension';
 import { showOutputChannel } from '../shell/output';
-import { TelemetryErrorEventNames } from '../telemetry';
+import { TelemetryErrorEventNames } from '../types/telemetryEventNames';
 import { refreshAllTreeViews, refreshResourcesTreeViews } from '../treeviews/treeViews';
 import { CommandId } from '../types/extensionIds';
 import { addKustomization } from './addKustomization';
