+++ src/commands/createKustomizationForPath.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { Uri, window, workspace } from 'vscode';
 import { failed } from '../types/errorable';
-import { getFolderGitInfo, getGitRepositoryforGitInfo } from '../git/gitInfo';
+import { getFolderGitInfo, getGitRepositoryforGitInfo } from '../cli/git/gitInfo';
 import { namespacedObject } from '../types/flux/object';
 import { getCurrentClusterInfo } from '../treeviews/treeViews';
 import { openConfigureGitOpsWebview } from '../webview-backend/configureGitOps/openWebview';
