+++ src/commands/fluxReconcileGitRepositoryForPath.ts
@@ -1,6 +1,6 @@
 import { Uri, window } from 'vscode';
-import { fluxTools } from '../flux/fluxTools';
-import { getFolderGitInfo, getGitRepositoryforGitInfo } from '../git/gitInfo';
+import { fluxTools } from '../cli/flux/fluxTools';
+import { getFolderGitInfo, getGitRepositoryforGitInfo } from '../cli/git/gitInfo';
 
 /**
  * Command to reconcile GitRepository for selected file
