+++ src/commands/pullGitRepository.ts
@@ -4,9 +4,9 @@ import semverSatisfies from 'semver/functions/satisfies';
 import safesh from 'shell-escape-tag';
 import { commands, Uri, window } from 'vscode';
 import { telemetry } from '../extension';
-import { checkGitVersion } from '../install';
+import { checkGitVersion } from '../cli/checkVersions';
 import { shell } from '../shell/shell';
-import { TelemetryErrorEventNames } from '../telemetry';
+import { TelemetryErrorEventNames } from '../types/telemetryEventNames';
 import { quoteFsPath } from '../utils/fsUtils';
 import { GitRepositoryNode } from '../treeviews/nodes/gitRepositoryNode';
