+++ src/commands/showInstalledVersions.ts
@@ -1,6 +1,6 @@
 import os from 'os';
 import { env, extensions, version, window } from 'vscode';
-import { getAzureVersion, getFluxVersion, getGitVersion, getKubectlVersion } from '../install';
+import { getAzureVersion, getFluxVersion, getGitVersion, getKubectlVersion } from '../cli/checkVersions';
 import { failed } from '../types/errorable';
 import { GitOpsExtensionConstants } from '../types/extensionIds';
