+++ src/commands/showLogs.ts
@@ -1,6 +1,6 @@
 
 import { commands, window } from 'vscode';
-import { kubernetesTools } from '../kubernetes/kubernetesTools';
+import { kubernetesTools } from '../cli/kubernetes/kubernetesTools';
 import { ClusterDeploymentNode } from '../treeviews/nodes/clusterDeploymentNode';
 import { ResourceNode, podResourceKind } from './showLogsTypes';
