+++ src/extension.ts
@@ -3,10 +3,12 @@ import { registerCommands } from './commands/commands';
 import { getExtensionVersion } from './commands/showInstalledVersions';
 import { showNewUserGuide } from './commands/showNewUserGuide';
 import { GlobalState, GlobalStateKey } from './globalState';
-import { checkFluxPrerequisites, checkWGEVersion, promptToInstallFlux } from './install';
+import { checkFluxPrerequisites, checkWGEVersion } from './cli/checkVersions';
+import { promptToInstallFlux } from './ui/promptToInstallFlux';
 import { shell } from './shell/shell';
-import { statusBar } from './statusBar';
-import { Telemetry, TelemetryEventNames } from './telemetry';
+import { statusBar } from './ui/statusBar';
+import { Telemetry } from './telemetry';
+import { TelemetryEventNames } from './types/telemetryEventNames';
 import { clusterTreeViewProvider, createTreeViews, sourceTreeViewProvider, workloadTreeViewProvider } from './treeviews/treeViews';
 import { succeeded } from './types/errorable';
 import { CommandId, ContextId, GitOpsExtensionConstants } from './types/extensionIds';
