+++ src/treeviews/dataProviders/clusterDataProvider.ts
@@ -1,8 +1,8 @@
 import { TreeItem, window } from 'vscode';
 import { setVSCodeContext } from '../../extension';
-import { fluxTools } from '../../flux/fluxTools';
-import { kubernetesTools } from '../../kubernetes/kubernetesTools';
-import { statusBar } from '../../statusBar';
+import { fluxTools } from '../../cli/flux/fluxTools';
+import { kubernetesTools } from '../../cli/kubernetes/kubernetesTools';
+import { statusBar } from '../../ui/statusBar';
 import { failed } from '../../types/errorable';
 import { ContextId } from '../../types/extensionIds';
 import { ClusterContextNode } from '../nodes/clusterContextNode';
