+++ src/treeviews/nodes/clusterContextNode.ts
@@ -1,7 +1,7 @@
 import { ExtensionMode, MarkdownString } from 'vscode';
 import { extensionContext, globalState, setVSCodeContext } from '../../extension';
-import { fluxVersion } from '../../install';
-import { currentContextName, kubernetesTools } from '../../kubernetes/kubernetesTools';
+import { fluxVersion } from '../../cli/checkVersions';
+import { currentContextName, kubernetesTools } from '../../cli/kubernetes/kubernetesTools';
 import { CommandId, ContextId } from '../../types/extensionIds';
 import { KubernetesCluster, KubernetesContextWithCluster } from '../../types/kubernetes/kubernetesConfig';
 import { ClusterProvider } from '../../types/kubernetes/kubernetesTypes';
