+++ src/treeviews/nodes/treeNode.ts
@@ -1,7 +1,7 @@
 import { KubernetesObject } from '@kubernetes/client-node';
 import { Command, MarkdownString, ThemeColor, ThemeIcon, TreeItem, TreeItemCollapsibleState, Uri } from 'vscode';
 import { asAbsolutePath } from '../../utils/asAbsolutePath';
-import { kubernetesTools } from '../../kubernetes/kubernetesTools';
+import { kubernetesTools } from '../../cli/kubernetes/kubernetesTools';
 import { CommandId } from '../../types/extensionIds';
 import { FileTypes } from '../../types/fileTypes';
 import { GitOpsTemplate } from '../../types/flux/gitOpsTemplate';
