+++ src/utils/jsonUtils.ts
@@ -1,6 +1,6 @@
 import { window } from 'vscode';
 import { telemetry } from '../extension';
-import { TelemetryErrorEventNames } from '../telemetry';
+import { TelemetryErrorEventNames } from '../types/telemetryEventNames';
 
 export function parseJson(jsonString: string): any {
 	let jsonData: any;
