+++ src/webview-backend/types.ts
@@ -1,4 +1,4 @@
-import { GitInfo } from '../git/gitInfo';
+import { GitInfo } from '../cli/git/gitInfo';
 import { GitOpsTemplate, TemplateParam } from '../types/flux/gitOpsTemplate';
 import { ClusterInfo, KubernetesObject } from '../types/kubernetes/kubernetesTypes';
